refactor(linx-impulse): change how productList loader work
