+++ linx-impulse/manifest.gen.ts
@@ -7,9 +7,10 @@ import * as $$$0 from "./loaders/config.ts";
 import * as $$$1 from "./loaders/products/hotsite.ts";
 import * as $$$2 from "./loaders/products/productList.ts";
 import * as $$$3 from "./loaders/products/productListingPage.ts";
-import * as $$$4 from "./loaders/search/autocomplete.ts";
-import * as $$$5 from "./loaders/search/popular.ts";
-import * as $$$6 from "./loaders/search/products.ts";
+import * as $$$4 from "./loaders/products/productLists.ts";
+import * as $$$5 from "./loaders/search/autocomplete.ts";
+import * as $$$6 from "./loaders/search/popular.ts";
+import * as $$$7 from "./loaders/search/products.ts";
 import * as $$$$$$0 from "./sections/Analytics/LinxImpulsePageView.tsx";
 import * as $$$$$$1 from "./sections/Script/LinxImpulseScript.tsx";
 
@@ -19,9 +20,10 @@ const manifest = {
     "linx-impulse/loaders/products/hotsite.ts": $$$1,
     "linx-impulse/loaders/products/productList.ts": $$$2,
     "linx-impulse/loaders/products/productListingPage.ts": $$$3,
-    "linx-impulse/loaders/search/autocomplete.ts": $$$4,
-    "linx-impulse/loaders/search/popular.ts": $$$5,
-    "linx-impulse/loaders/search/products.ts": $$$6,
+    "linx-impulse/loaders/products/productLists.ts": $$$4,
+    "linx-impulse/loaders/search/autocomplete.ts": $$$5,
+    "linx-impulse/loaders/search/popular.ts": $$$6,
+    "linx-impulse/loaders/search/products.ts": $$$7,
   },
   "sections": {
     "linx-impulse/sections/Analytics/LinxImpulsePageView.tsx": $$$$$$0,
